+++ CHANGELOG.md
@@ -5,7 +5,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased] 
-- Updated API-Builder to  version Zagreb
+### Added
+- Custom-Properties for an API now looked up
+  - Index mapping now includes all configured custom properties
+
+### Changed
+- Updated API-Builder to version Agra
+- Elasticsearch configuration now managed by API-Builder project and no longer by Logstash
+  - Index-Templates
+  - ILM-Policies has been added
+  - Rollup-Jobs has been added
 
 ## [1.0.0] 2020-10-01
 ### Added